fix off by one errors and signed/unsigned comparison errors

+++ resample.h
@@ -56,10 +56,10 @@
 struct rs_data {
 	double factor;
 	unsigned int time;
-	int in_buf_ptr;
-	int out_buf_ptr;
-	int in_buf_read;
-	int in_buf_offset;
+	unsigned short in_buf_ptr;
+	unsigned short out_buf_ptr;
+	unsigned short in_buf_read;
+	unsigned short in_buf_offset;
 	int in_buf_used;
 	int in_buf_size;
 	int out_buf_size;
